moved UUID class into binary file
